fix: remaining audit findings M4, M8, L2, L3
M4: validateMint now uses PI controller's dynamic redemptionPrice
     instead of static HEAT_LAUNCH_RATIO. User mints and protocol
     mints now use the same rate.

M8: LP fee tracking now uses actual fee (outputNoFee - outputAmount)
     from constant-product formula instead of linear approximation.
     Prevents fee accounting drift over time.

L2: calculateCdInterest uses __uint128_t intermediate to prevent
     overflow for large deposits and epoch rates.

L3: CD reserve cap now uses pre-mint pool size instead of
     post-decrement value, preventing premature reserve drain to treasury.

popTransaction fee reversal also updated to use actual fee (M8).

+++ src/CryptoNoteCore/Currency.cpp
@@ -333,9 +333,8 @@ double Currency::getBurnPercentage() const {
     uint64_t interest = 0;
     for (uint64_t e = startEpoch; e <= endEpoch && e < epochCount; ++e) {
       uint64_t epochRate = commitmentIndex.getEpochFeeRate(e);
-      // interest += amount * epochRate / RATE_PRECISION
-      // max: 8e9 * 1e6 = 8e15 < 2^63, safe in 64-bit
-      interest += (amount * epochRate) / parameters::FEE_POOL_RATE_PRECISION;
+      // Use 128-bit intermediate to prevent overflow for large deposits
+      interest += (uint64_t)(((__uint128_t)amount * epochRate) / parameters::FEE_POOL_RATE_PRECISION);
     }
 
     return interest;
